+++ base/src/main/java/proguard/analysis/cpa/jvm/domain/taint/JvmTaintCpa.java
@@ -31,6 +31,8 @@ import proguard.analysis.cpa.defaults.StopJoinOperator;
 import proguard.analysis.cpa.domain.taint.TaintSource;
 import proguard.analysis.cpa.interfaces.AbstractDomain;
 import proguard.analysis.cpa.jvm.state.JvmAbstractState;
+import proguard.analysis.cpa.jvm.witness.JvmMemoryLocation;
+import proguard.analysis.datastructure.callgraph.Call;
 import proguard.classfile.MethodSignature;
 import proguard.classfile.Signature;
 
@@ -49,6 +51,7 @@ public class JvmTaintCpa extends SimpleCpa {
     this(
         createSourcesMap(sources),
         new DelegateAbstractDomain<JvmAbstractState<SetAbstractState<JvmTaintSource>>>(),
+        Collections.emptyMap(),
         Collections.emptyMap());
   }
 
@@ -61,11 +64,13 @@ public class JvmTaintCpa extends SimpleCpa {
    */
   public JvmTaintCpa(
       Set<? extends JvmTaintSource> sources,
-      Map<MethodSignature, JvmTaintTransformer> taintTransformers) {
+      Map<MethodSignature, JvmTaintTransformer> taintTransformers,
+      Map<Call, Set<JvmMemoryLocation>> extraTaintPropagationLocations) {
     this(
         createSourcesMap(sources),
         new DelegateAbstractDomain<JvmAbstractState<SetAbstractState<JvmTaintSource>>>(),
-        taintTransformers);
+        taintTransformers,
+        extraTaintPropagationLocations);
   }
 
   /**
@@ -81,16 +86,18 @@ public class JvmTaintCpa extends SimpleCpa {
     this(
         signaturesToSources,
         new DelegateAbstractDomain<JvmAbstractState<SetAbstractState<JvmTaintSource>>>(),
-        taintTransformers);
+        taintTransformers,
+        Collections.emptyMap());
   }
 
   private JvmTaintCpa(
       Map<Signature, Set<JvmTaintSource>> sources,
       AbstractDomain abstractDomain,
-      Map<MethodSignature, JvmTaintTransformer> taintTransformers) {
+      Map<MethodSignature, JvmTaintTransformer> taintTransformers,
+      Map<Call, Set<JvmMemoryLocation>> extraTaintPropagationLocations) {
     super(
         abstractDomain,
-        new JvmTaintTransferRelation(sources, taintTransformers),
+        new JvmTaintTransferRelation(sources, taintTransformers, extraTaintPropagationLocations),
         new MergeJoinOperator(abstractDomain),
         new StopJoinOperator(abstractDomain));
   }